+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog for Cass Operator, new PRs should update the `main / unreleased` secti
 
 ## unreleased
 
+* [CHANGE] [#875](https://github.com/k8ssandra/cass-operator/issues/875) Update to Go 1.25, Kubernetes 1.34 and UBI10 as the base image. Also, Vector updates and other smaller dependency updates.
+* [CHANGE] [#876](https://github.com/k8ssandra/cass-operator/issues/876) Drop support for ImageConfig v1beta1, the only supported version from now on is v1beta2
 * [CHANGE] [#865](https://github.com/k8ssandra/cass-operator/issues/865) Add VolumeMount for the management-api-server-certs-volume volume to all containers instead of only cassandra container
 * [CHANGE] [#850](https://github.com/k8ssandra/cass-operator/issues/850) The default retry policy for CassandraTask has changed to OnFailure
 * [CHANGE] [#775](https://github.com/k8ssandra/cass-operator/issues/775) Set generic podFilter as default filter for CassandraTask
@@ -23,6 +25,7 @@ Changelog for Cass Operator, new PRs should update the `main / unreleased` secti
 
 * [ENHANCEMENT] [#868](https://github.com/k8ssandra/cass-operator/issues/868) Add new flag to the controller, --metrics-secure-auth to enable/disable the authentication/authorization of metrics endpoint. Default is disabled (TLS is still enabled by default)
 * [BUGFIX] [#862](https://github.com/k8ssandra/cass-operator/issues/862) If volumeMount was provided in PodTemplateSpec and AdditionalVolumes, the latter would override the PodTemplateSpec one. This was unintentional, the PodTemplateSpec selection should be the final one.
+* [BUGFIX] [#870](https://github.com/k8ssandra/cass-operator/issues/870) Fix Kustomize deployment to properly add the path to the TLS cert for the webhooks. Remove OperConfig and ImageConfig v1
 
 ## v1.28.0